+++ src/test/integrationExplorer.test.ts
@@ -1,3 +1,19 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 import * as vscode from 'vscode';
 import * as chai from 'chai';
 import { CamelKNodeProvider, TreeNode } from '../CamelKNodeProvider';
@@ -14,20 +30,32 @@ suite('Camel-k Integrations View', () => {
 
 	setup(() => {
 		sandbox = sinon.createSandbox();
-
 		integrationExplorer = new CamelKNodeProvider();
+		integrationExplorer.setRetrieveIntegrations(false);
 	});
 
 	teardown(() => {
 		sandbox.restore();
 	});
 
-	test('getChildren call with dummy node should add integration to tree data model', async () => {
+	test('adding a single child should trigger a refresh', async () => {
+		integrationExplorer.resetList();
+		const refreshStub = sandbox.stub(integrationExplorer, 'refresh');
+		var children = await integrationExplorer.getChildren();
 		const newNode = new TreeNode("string", "mockIntegration", vscode.TreeItemCollapsibleState.None);
-        const children = await integrationExplorer.getChildren(newNode);
-
-        expect(children.length).equals(1);
-        expect(children[0].label).equals("mockIntegration");
+		integrationExplorer.addChild(children, newNode, false);
+		expect(children.length).equals(1);
+		expect(children[0].label).equals("mockIntegration");
+		expect(refreshStub).calledOnce;
 	});
 
+	test('adding and removing a child should trigger refresh twice', async () => {
+		integrationExplorer.resetList();
+		const refreshStub = sandbox.stub(integrationExplorer, 'refresh');
+		var children = await integrationExplorer.getChildren();
+		const newNode = new TreeNode("string", "mockIntegration", vscode.TreeItemCollapsibleState.None);
+		integrationExplorer.addChild(children, newNode);
+		integrationExplorer.removeChild(children, newNode);
+		expect(refreshStub).calledTwice;
+	});
 });